Takes in all keystrokes now, including arrows keys. Rotating but degrades with each rotation.

+++ rocketship.c
@@ -57,8 +57,8 @@ void rotate(int right, int left){
     int x = rocketship.points[i].x;
     int y = rocketship.points[i].y;
 
-    double rotated_x = x * cos(rotate_radian) - y * sin(rotate_radian);
-    double rotated_y = x * sin(rotate_radian) + y * cos(rotate_radian);
+    int rotated_x = x * cos(rotate_radian) - y * sin(rotate_radian);
+    int rotated_y = x * sin(rotate_radian) + y * cos(rotate_radian);
     rocketship.points[i].x = rotated_x;
     rocketship.points[i].y = rotated_y;
     fprintf(stderr, "B%dA%d ", rotate_radian, rotated_x);
